// src/app/(root)/forum-prestacao-contas/interface.ts
export interface Municipio {
  id: number;
  provincia_id: number;
  nome: string;
  deleted_at: string | null;
}
// ...
export interface Usuario {
  id: number;
  nome: string;
  email: string;
  telefone: string;
  bi: string;
  data_nascimento: string; // "YYYY-MM-DD"
  genero: string;
  status: string;
  tipo: string;
  natural_id: number | null;
  criado_por: number | null;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
}
// ...
export interface Forum {
  id: number
  tipo: string
  titulo: string
  descricao: string | null
  data_inicio: string
  data_fim: string
  local: string | null
  municipio_id: number
  criado_por: Usuario
  created_at: string
  updated_at: string
  deleted_at: string | null
  municipio: Municipio
  participantes: Participante[]
  forun_contribuicao: Contribuicao[]
  forun_prestacao_contas: Contribuicao[]
}
// ...
export interface Participante {
  // Adicione campos reais se houver estrutura definida
  nome: string
  papel: string
}

export interface Contribuicao {
  id: number
  projeto: Proposta
  prioridade: string
  criado_por: number
  created_at: string
}